refactor(5591): migrate import-srp component to React Router v6 compatibility
