+++ ui/pages/multi-srp/import-srp/import-srp.test.tsx
@@ -7,7 +7,7 @@ import {
 } from '@testing-library/react';
 import configureMockStore from 'redux-mock-store';
 import thunk from 'redux-thunk';
-import { renderWithProvider } from '../../../../test/jest/rendering';
+import { renderWithProvider } from '../../../../test/lib/render-helpers-navigate';
 import mockState from '../../../../test/data/mock-state.json';
 import { importMnemonicToVault } from '../../../store/actions';
 import { ImportSrp } from './import-srp';
