+++ ui/pages/multi-srp/import-srp/import-srp.tsx
@@ -8,7 +8,7 @@ import React, {
 import { useDispatch, useSelector } from 'react-redux';
 import { isValidMnemonic } from '@ethersproject/hdnode';
 import { wordlist } from '@metamask/scure-bip39/dist/wordlists/english';
-import { useHistory } from 'react-router-dom';
+import { useNavigate } from 'react-router-dom-v5-compat';
 import { useI18nContext } from '../../../hooks/useI18nContext';
 import {
   hideWarning,
@@ -64,7 +64,7 @@ const defaultNumberOfWords = 12;
 
 export const ImportSrp = () => {
   const t = useI18nContext();
-  const history = useHistory();
+  const navigate = useNavigate();
   const trackEvent = useContext(MetaMetricsContext);
   const dispatch = useDispatch<MetaMaskReduxDispatch>();
   const [srpError, setSrpError] = useState('');
@@ -133,7 +133,7 @@ export const ImportSrp = () => {
       });
     }
 
-    history.push(DEFAULT_ROUTE);
+    navigate(DEFAULT_ROUTE);
     dispatch(setShowNewSrpAddedToast(true));
   }
 
@@ -322,7 +322,7 @@ export const ImportSrp = () => {
             ariaLabel="back"
             iconName={IconName.ArrowLeft}
             onClick={() => {
-              history.push(DEFAULT_ROUTE);
+              navigate(DEFAULT_ROUTE);
             }}
           />
         }
@@ -331,7 +331,7 @@ export const ImportSrp = () => {
             ariaLabel="close"
             iconName={IconName.Close}
             onClick={() => {
-              history.push(DEFAULT_ROUTE);
+              navigate(DEFAULT_ROUTE);
             }}
           />
         }
